backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Django==5.2.3
djangorestframework==3.16.0
dnspython==2.7.0
flasgger==0.9.7.1
Flask==3.1.1
Flask-PyMongo==3.0.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mistune==3.1.3
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pymongo==4.13.2
pytest==8.4.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
